requirements.txt
# LangChain et dÃ©pendances
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-ollama>=0.1.0

# Scikit-learn pour vector store
scikit-learn>=1.3.0

# Interface web
flask>=3.0.0

# Parsing de documents
pypdf>=3.0.0
python-docx>=1.1.0

# Utilitaires
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
